steg_studio/requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
cryptography>=41.0.0
numpy>=1.24.0
pyaudio>=0.2.14
